Ahhhh

+++ testedememoria/matrixmulti.cl
@@ -1,4 +1,4 @@
-__kernel void matrixmulti(__global int* a, __global int* b)
+__kernel void matrixmulti(__global float* a, __global float* b)
 {
     unsigned int row = get_global_id(0);
     unsigned int column = get_global_id(1);
